requirements.txt
Pillow
numpy
pydub
opencv-python
click